bindinfo: migrate test-infra to testify(staging-3)
