+++ bindinfo/bind_test.go
@@ -41,7 +41,6 @@ import (
 	"github.com/pingcap/tidb/util"
 	"github.com/pingcap/tidb/util/logutil"
 	utilparser "github.com/pingcap/tidb/util/parser"
-	"github.com/pingcap/tidb/util/stmtsummary"
 	"github.com/pingcap/tidb/util/testkit"
 	"github.com/pingcap/tidb/util/testleak"
 	"github.com/tikv/client-go/v2/testutils"
@@ -518,175 +517,6 @@ func (s *testSuite) TestErrorBind(c *C) {
 	c.Check(chk.NumRows(), Equals, 0)
 }
 
-func (s *testSuite) TestDMLCapturePlanBaseline(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	s.cleanBindingEnv(tk)
-	stmtsummary.StmtSummaryByDigestMap.Clear()
-	tk.MustExec(" set @@tidb_capture_plan_baselines = on")
-	defer func() {
-		tk.MustExec(" set @@tidb_capture_plan_baselines = off")
-	}()
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t(a int, b int, c int, key idx_b(b), key idx_c(c))")
-	tk.MustExec("create table t1 like t")
-	s.domain.BindHandle().CaptureBaselines()
-	tk.MustQuery("show global bindings").Check(testkit.Rows())
-	tk.MustExec("delete from t where b = 1 and c > 1")
-	tk.MustExec("delete from t where b = 1 and c > 1")
-	tk.MustExec("update t set a = 1 where b = 1 and c > 1")
-	tk.MustExec("update t set a = 1 where b = 1 and c > 1")
-	tk.MustExec("insert into t1 select * from t where t.b = 1 and t.c > 1")
-	tk.MustExec("insert into t1 select * from t where t.b = 1 and t.c > 1")
-	tk.MustExec("replace into t1 select * from t where t.b = 1 and t.c > 1")
-	tk.MustExec("replace into t1 select * from t where t.b = 1 and t.c > 1")
-	tk.MustExec("insert into t1 values(1,1,1)")
-	tk.MustExec("insert into t1 values(1,1,1)")
-	tk.MustExec("replace into t1 values(1,1,1)")
-	tk.MustExec("replace into t1 values(1,1,1)")
-	tk.MustExec("admin capture bindings")
-	rows := tk.MustQuery("show global bindings").Rows()
-	c.Assert(len(rows), Equals, 0)
-
-	c.Assert(tk.Se.Auth(&auth.UserIdentity{Username: "root", Hostname: "%"}, nil, nil), IsTrue)
-	tk.MustExec("delete from t where b = 1 and c > 1")
-	tk.MustExec("delete from t where b = 1 and c > 1")
-	tk.MustExec("update t set a = 1 where b = 1 and c > 1")
-	tk.MustExec("update t set a = 1 where b = 1 and c > 1")
-	tk.MustExec("insert into t1 select * from t where t.b = 1 and t.c > 1")
-	tk.MustExec("insert into t1 select * from t where t.b = 1 and t.c > 1")
-	tk.MustExec("replace into t1 select * from t where t.b = 1 and t.c > 1")
-	tk.MustExec("replace into t1 select * from t where t.b = 1 and t.c > 1")
-	tk.MustExec("insert into t1 values(1,1,1)")
-	tk.MustExec("insert into t1 values(1,1,1)")
-	tk.MustExec("replace into t1 values(1,1,1)")
-	tk.MustExec("replace into t1 values(1,1,1)")
-	tk.MustExec("admin capture bindings")
-	rows = tk.MustQuery("show global bindings").Sort().Rows()
-	c.Assert(len(rows), Equals, 4)
-	c.Assert(rows[0][0], Equals, "delete from `test` . `t` where `b` = ? and `c` > ?")
-	c.Assert(rows[0][1], Equals, "DELETE /*+ use_index(@`del_1` `test`.`t` `idx_b`)*/ FROM `test`.`t` WHERE `b` = 1 AND `c` > 1")
-	c.Assert(rows[1][0], Equals, "insert into `test` . `t1` select * from `test` . `t` where `t` . `b` = ? and `t` . `c` > ?")
-	c.Assert(rows[1][1], Equals, "INSERT INTO `test`.`t1` SELECT /*+ use_index(@`sel_1` `test`.`t` `idx_b`)*/ * FROM `test`.`t` WHERE `t`.`b` = 1 AND `t`.`c` > 1")
-	c.Assert(rows[2][0], Equals, "replace into `test` . `t1` select * from `test` . `t` where `t` . `b` = ? and `t` . `c` > ?")
-	c.Assert(rows[2][1], Equals, "REPLACE INTO `test`.`t1` SELECT /*+ use_index(@`sel_1` `test`.`t` `idx_b`)*/ * FROM `test`.`t` WHERE `t`.`b` = 1 AND `t`.`c` > 1")
-	c.Assert(rows[3][0], Equals, "update `test` . `t` set `a` = ? where `b` = ? and `c` > ?")
-	c.Assert(rows[3][1], Equals, "UPDATE /*+ use_index(@`upd_1` `test`.`t` `idx_b`)*/ `test`.`t` SET `a`=1 WHERE `b` = 1 AND `c` > 1")
-}
-
-func (s *testSuite) TestCapturePlanBaseline(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	s.cleanBindingEnv(tk)
-	stmtsummary.StmtSummaryByDigestMap.Clear()
-	tk.MustExec(" set @@tidb_capture_plan_baselines = on")
-	defer func() {
-		tk.MustExec(" set @@tidb_capture_plan_baselines = off")
-	}()
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t(a int)")
-	s.domain.BindHandle().CaptureBaselines()
-	tk.MustQuery("show global bindings").Check(testkit.Rows())
-	tk.MustExec("select count(*) from t where a > 10")
-	tk.MustExec("select count(*) from t where a > 10")
-	tk.MustExec("admin capture bindings")
-	rows := tk.MustQuery("show global bindings").Rows()
-	c.Assert(len(rows), Equals, 0)
-
-	c.Assert(tk.Se.Auth(&auth.UserIdentity{Username: "root", Hostname: "%"}, nil, nil), IsTrue)
-	tk.MustExec("select * from t where a > 10")
-	tk.MustExec("select * from t where a > 10")
-	tk.MustExec("admin capture bindings")
-	rows = tk.MustQuery("show global bindings").Rows()
-	c.Assert(len(rows), Equals, 1)
-	c.Assert(rows[0][0], Equals, "select * from `test` . `t` where `a` > ?")
-	c.Assert(rows[0][1], Equals, "SELECT /*+ use_index(@`sel_1` `test`.`t` )*/ * FROM `test`.`t` WHERE `a` > 10")
-}
-
-func (s *testSuite) TestCaptureDBCaseSensitivity(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	s.cleanBindingEnv(tk)
-	stmtsummary.StmtSummaryByDigestMap.Clear()
-	tk.MustExec("drop database if exists SPM")
-	tk.MustExec("create database SPM")
-	tk.MustExec("use SPM")
-	tk.MustExec("create table t(a int, b int, key(b))")
-	tk.MustExec("create global binding for select * from t using select /*+ use_index(t) */ * from t")
-	c.Assert(tk.Se.Auth(&auth.UserIdentity{Username: "root", Hostname: "%"}, nil, nil), IsTrue)
-	tk.MustExec("select /*+ use_index(t,b) */ * from t")
-	tk.MustExec("select /*+ use_index(t,b) */ * from t")
-	tk.MustExec("admin capture bindings")
-	// The capture should ignore the case sensitivity for DB name when checking if any binding exists,
-	// so there would be no new binding captured.
-	rows := tk.MustQuery("show global bindings").Rows()
-	c.Assert(len(rows), Equals, 1)
-	c.Assert(rows[0][1], Equals, "SELECT /*+ use_index(`t` )*/ * FROM `SPM`.`t`")
-	c.Assert(rows[0][8], Equals, "manual")
-}
-
-func (s *testSuite) TestCaptureBaselinesDefaultDB(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	s.cleanBindingEnv(tk)
-	stmtsummary.StmtSummaryByDigestMap.Clear()
-	tk.MustExec(" set @@tidb_capture_plan_baselines = on")
-	defer func() {
-		tk.MustExec(" set @@tidb_capture_plan_baselines = off")
-	}()
-	tk.MustExec("use test")
-	tk.MustExec("drop database if exists spm")
-	tk.MustExec("create database spm")
-	tk.MustExec("create table spm.t(a int, index idx_a(a))")
-	c.Assert(tk.Se.Auth(&auth.UserIdentity{Username: "root", Hostname: "%"}, nil, nil), IsTrue)
-	tk.MustExec("select * from spm.t ignore index(idx_a) where a > 10")
-	tk.MustExec("select * from spm.t ignore index(idx_a) where a > 10")
-	tk.MustExec("admin capture bindings")
-	rows := tk.MustQuery("show global bindings").Rows()
-	c.Assert(len(rows), Equals, 1)
-	// Default DB should be "" when all columns have explicit database name.
-	c.Assert(rows[0][2], Equals, "")
-	c.Assert(rows[0][3], Equals, "using")
-	tk.MustExec("use spm")
-	tk.MustExec("select * from spm.t where a > 10")
-	// Should use TableScan because of the "ignore index" binding.
-	c.Assert(len(tk.Se.GetSessionVars().StmtCtx.IndexNames), Equals, 0)
-}
-
-func (s *testSuite) TestCapturePreparedStmt(c *C) {
-	originalVal := config.CheckTableBeforeDrop
-	config.CheckTableBeforeDrop = true
-	defer func() {
-		config.CheckTableBeforeDrop = originalVal
-	}()
-
-	tk := testkit.NewTestKit(c, s.store)
-	s.cleanBindingEnv(tk)
-	stmtsummary.StmtSummaryByDigestMap.Clear()
-	c.Assert(tk.Se.Auth(&auth.UserIdentity{Username: "root", Hostname: "%"}, nil, nil), IsTrue)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t(a int, b int, c int, key idx_b(b), key idx_c(c))")
-	c.Assert(tk.MustUseIndex("select * from t where b = 1 and c > 1", "idx_b(b)"), IsTrue)
-	tk.MustExec("prepare stmt from 'select /*+ use_index(t,idx_c) */ * from t where b = ? and c > ?'")
-	tk.MustExec("set @p = 1")
-	tk.MustExec("execute stmt using @p, @p")
-	tk.MustExec("execute stmt using @p, @p")
-
-	tk.MustQuery("show global bindings").Check(testkit.Rows())
-	tk.MustExec("admin capture bindings")
-	rows := tk.MustQuery("show global bindings").Rows()
-	c.Assert(len(rows), Equals, 1)
-	c.Assert(rows[0][0], Equals, "select * from `test` . `t` where `b` = ? and `c` > ?")
-	c.Assert(rows[0][1], Equals, "SELECT /*+ use_index(@`sel_1` `test`.`t` `idx_c`)*/ * FROM `test`.`t` WHERE `b` = ? AND `c` > ?")
-
-	c.Assert(tk.MustUseIndex("select /*+ use_index(t,idx_b) */ * from t where b = 1 and c > 1", "idx_c(c)"), IsTrue)
-	tk.MustExec("admin flush bindings")
-	tk.MustExec("admin evolve bindings")
-	rows = tk.MustQuery("show global bindings").Rows()
-	c.Assert(len(rows), Equals, 1)
-	c.Assert(rows[0][0], Equals, "select * from `test` . `t` where `b` = ? and `c` > ?")
-	c.Assert(rows[0][1], Equals, "SELECT /*+ use_index(@`sel_1` `test`.`t` `idx_c`)*/ * FROM `test`.`t` WHERE `b` = ? AND `c` > ?")
-}
-
 func (s *testSuite) TestDMLEvolveBaselines(c *C) {
 	originalVal := config.CheckTableBeforeDrop
 	config.CheckTableBeforeDrop = true
@@ -992,42 +822,6 @@ func (s *testSuite) TestHintsSetID(c *C) {
 	c.Assert(bind.ID, Equals, "")
 }
 
-func (s *testSuite) TestCapturePlanBaselineIgnoreTiFlash(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	s.cleanBindingEnv(tk)
-	stmtsummary.StmtSummaryByDigestMap.Clear()
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t(a int, b int, key(a), key(b))")
-	c.Assert(tk.Se.Auth(&auth.UserIdentity{Username: "root", Hostname: "%"}, nil, nil), IsTrue)
-	tk.MustExec("select * from t")
-	tk.MustExec("select * from t")
-	// Create virtual tiflash replica info.
-	dom := domain.GetDomain(tk.Se)
-	is := dom.InfoSchema()
-	db, exists := is.SchemaByName(model.NewCIStr("test"))
-	c.Assert(exists, IsTrue)
-	for _, tblInfo := range db.Tables {
-		if tblInfo.Name.L == "t" {
-			tblInfo.TiFlashReplica = &model.TiFlashReplicaInfo{
-				Count:     1,
-				Available: true,
-			}
-		}
-	}
-	// Here the plan is the TiFlash plan.
-	rows := tk.MustQuery("explain select * from t").Rows()
-	c.Assert(fmt.Sprintf("%v", rows[len(rows)-1][2]), Equals, "cop[tiflash]")
-
-	tk.MustQuery("show global bindings").Check(testkit.Rows())
-	tk.MustExec("admin capture bindings")
-	// Don't have the TiFlash plan even we have TiFlash replica.
-	rows = tk.MustQuery("show global bindings").Rows()
-	c.Assert(len(rows), Equals, 1)
-	c.Assert(rows[0][0], Equals, "select * from `test` . `t`")
-	c.Assert(rows[0][1], Equals, "SELECT /*+ use_index(@`sel_1` `test`.`t` )*/ * FROM `test`.`t`")
-}
-
 func (s *testSuite) TestNotEvolvePlanForReadStorageHint(c *C) {
 	originalVal := config.CheckTableBeforeDrop
 	config.CheckTableBeforeDrop = true
@@ -1174,64 +968,6 @@ func (s *testSuite) TestInvisibleIndex(c *C) {
 	tk.MustExec("drop binding for select * from t")
 }
 
-func (s *testSuite) TestBindingSource(c *C) {
-	originalVal := config.CheckTableBeforeDrop
-	config.CheckTableBeforeDrop = true
-	defer func() {
-		config.CheckTableBeforeDrop = originalVal
-	}()
-
-	tk := testkit.NewTestKit(c, s.store)
-	s.cleanBindingEnv(tk)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t(a int, index idx_a(a))")
-
-	// Test Source for SQL created sql
-	tk.MustExec("create global binding for select * from t where a > 10 using select * from t ignore index(idx_a) where a > 10")
-	bindHandle := s.domain.BindHandle()
-	sql, hash := normalizeWithDefaultDB(c, "select * from t where a > ?", "test")
-	bindData := bindHandle.GetBindRecord(hash, sql, "test")
-	c.Check(bindData, NotNil)
-	c.Check(bindData.OriginalSQL, Equals, "select * from `test` . `t` where `a` > ?")
-	c.Assert(len(bindData.Bindings), Equals, 1)
-	bind := bindData.Bindings[0]
-	c.Assert(bind.Source, Equals, bindinfo.Manual)
-
-	// Test Source for evolved sql
-	tk.MustExec("set @@tidb_evolve_plan_baselines=1")
-	tk.MustQuery("select * from t where a > 10")
-	bindHandle.SaveEvolveTasksToStore()
-	sql, hash = normalizeWithDefaultDB(c, "select * from t where a > ?", "test")
-	bindData = bindHandle.GetBindRecord(hash, sql, "test")
-	c.Check(bindData, NotNil)
-	c.Check(bindData.OriginalSQL, Equals, "select * from `test` . `t` where `a` > ?")
-	c.Assert(len(bindData.Bindings), Equals, 2)
-	bind = bindData.Bindings[1]
-	c.Assert(bind.Source, Equals, bindinfo.Evolve)
-	tk.MustExec("set @@tidb_evolve_plan_baselines=0")
-
-	// Test Source for captured sqls
-	stmtsummary.StmtSummaryByDigestMap.Clear()
-	tk.MustExec("set @@tidb_capture_plan_baselines = on")
-	defer func() {
-		tk.MustExec("set @@tidb_capture_plan_baselines = off")
-	}()
-	tk.MustExec("use test")
-	c.Assert(tk.Se.Auth(&auth.UserIdentity{Username: "root", Hostname: "%"}, nil, nil), IsTrue)
-	tk.MustExec("select * from t ignore index(idx_a) where a < 10")
-	tk.MustExec("select * from t ignore index(idx_a) where a < 10")
-	tk.MustExec("admin capture bindings")
-	bindHandle.CaptureBaselines()
-	sql, hash = normalizeWithDefaultDB(c, "select * from t where a < ?", "test")
-	bindData = bindHandle.GetBindRecord(hash, sql, "test")
-	c.Check(bindData, NotNil)
-	c.Check(bindData.OriginalSQL, Equals, "select * from `test` . `t` where `a` < ?")
-	c.Assert(len(bindData.Bindings), Equals, 1)
-	bind = bindData.Bindings[0]
-	c.Assert(bind.Source, Equals, bindinfo.Capture)
-}
-
 func (s *testSuite) TestSPMHitInfo(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	s.cleanBindingEnv(tk)
@@ -1317,147 +1053,6 @@ func (s *testSuite) TestDMLIndexHintBind(c *C) {
 	c.Assert(tk.MustUseIndex("delete from t where b = 1 and c > 1", "idx_c(c)"), IsTrue)
 }
 
-func (s *testSuite) TestCapturedBindingCharset(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	s.cleanBindingEnv(tk)
-	stmtsummary.StmtSummaryByDigestMap.Clear()
-	c.Assert(tk.Se.Auth(&auth.UserIdentity{Username: "root", Hostname: "%"}, nil, nil), IsTrue)
-	tk.MustExec("use test")
-	tk.MustExec("create table t(name varchar(25), index idx(name))")
-
-	tk.MustExec("set character_set_connection = 'ascii'")
-	tk.MustExec("update t set name = 'hello' where name <= 'abc'")
-	tk.MustExec("update t set name = 'hello' where name <= 'abc'")
-	tk.MustExec("set character_set_connection = 'utf8mb4'")
-	tk.MustExec("admin capture bindings")
-	rows := tk.MustQuery("show global bindings").Rows()
-	c.Assert(len(rows), Equals, 1)
-	c.Assert(rows[0][0], Equals, "update `test` . `t` set `name` = ? where `name` <= ?")
-	c.Assert(rows[0][1], Equals, "UPDATE /*+ use_index(@`upd_1` `test`.`t` `idx`)*/ `test`.`t` SET `name`='hello' WHERE `name` <= 'abc'")
-	// Charset and Collation are empty now, they are not used currently.
-	c.Assert(rows[0][6], Equals, "")
-	c.Assert(rows[0][7], Equals, "")
-}
-
-func (s *testSuite) TestConcurrentCapture(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	s.cleanBindingEnv(tk)
-	// Simulate an existing binding generated by concurrent CREATE BINDING, which has not been synchronized to current tidb-server yet.
-	// Actually, it is more common to be generated by concurrent baseline capture, I use Manual just for simpler test verification.
-	tk.MustExec("insert into mysql.bind_info values('select * from `test` . `t`', 'select * from `test` . `t`', '', 'using', '2000-01-01 09:00:00', '2000-01-01 09:00:00', '', '','" +
-		bindinfo.Manual + "')")
-	tk.MustQuery("select original_sql, source from mysql.bind_info where source != 'builtin'").Check(testkit.Rows(
-		"select * from `test` . `t` manual",
-	))
-	stmtsummary.StmtSummaryByDigestMap.Clear()
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t(a int, b int)")
-	c.Assert(tk.Se.Auth(&auth.UserIdentity{Username: "root", Hostname: "%"}, nil, nil), IsTrue)
-	tk.MustExec("select * from t")
-	tk.MustExec("select * from t")
-	tk.MustExec("admin capture bindings")
-	tk.MustQuery("select original_sql, source, status from mysql.bind_info where source != 'builtin'").Check(testkit.Rows(
-		"select * from `test` . `t` manual deleted",
-		"select * from `test` . `t` capture using",
-	))
-}
-
-func (s *testSuite) TestUpdateSubqueryCapture(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	s.cleanBindingEnv(tk)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t1, t2")
-	tk.MustExec("create table t1(a int, b int, c int, key idx_b(b))")
-	tk.MustExec("create table t2(a int, b int)")
-	stmtsummary.StmtSummaryByDigestMap.Clear()
-	c.Assert(tk.Se.Auth(&auth.UserIdentity{Username: "root", Hostname: "%"}, nil, nil), IsTrue)
-	tk.MustExec("update t1 set b = 1 where b = 2 and (a in (select a from t2 where b = 1) or c in (select a from t2 where b = 1))")
-	tk.MustExec("update t1 set b = 1 where b = 2 and (a in (select a from t2 where b = 1) or c in (select a from t2 where b = 1))")
-	tk.MustExec("admin capture bindings")
-	rows := tk.MustQuery("show global bindings").Rows()
-	c.Assert(len(rows), Equals, 1)
-	bindSQL := "UPDATE /*+ use_index(@`upd_1` `test`.`t1` `idx_b`), use_index(@`sel_1` `test`.`t2` ), hash_join(@`upd_1` `test`.`t1`), use_index(@`sel_2` `test`.`t2` )*/ `test`.`t1` SET `b`=1 WHERE `b` = 2 AND (`a` IN (SELECT `a` FROM `test`.`t2` WHERE `b` = 1) OR `c` IN (SELECT `a` FROM `test`.`t2` WHERE `b` = 1))"
-	c.Assert(rows[0][1], Equals, bindSQL)
-	tk.MustExec(bindSQL)
-	c.Assert(tk.Se.GetSessionVars().StmtCtx.GetWarnings(), HasLen, 0)
-}
-
-func (s *testSuite) TestIssue20417(c *C) {
-	originalVal := config.CheckTableBeforeDrop
-	config.CheckTableBeforeDrop = true
-	defer func() {
-		config.CheckTableBeforeDrop = originalVal
-	}()
-
-	tk := testkit.NewTestKit(c, s.store)
-	s.cleanBindingEnv(tk)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec(`CREATE TABLE t (
-		 pk VARBINARY(36) NOT NULL PRIMARY KEY,
-		 b BIGINT NOT NULL,
-		 c BIGINT NOT NULL,
-		 pad VARBINARY(2048),
-		 INDEX idxb(b),
-		 INDEX idxc(c)
-		)`)
-
-	// Test for create binding
-	s.cleanBindingEnv(tk)
-	tk.MustExec("create global binding for select * from t using select /*+ use_index(t, idxb) */ * from t")
-	rows := tk.MustQuery("show global bindings").Rows()
-	c.Assert(len(rows), Equals, 1)
-	c.Assert(rows[0][0], Equals, "select * from `test` . `t`")
-	c.Assert(rows[0][1], Equals, "SELECT /*+ use_index(`t` `idxb`)*/ * FROM `test`.`t`")
-	c.Assert(tk.MustUseIndex("select * from t", "idxb(b)"), IsTrue)
-	c.Assert(tk.MustUseIndex("select * from test.t", "idxb(b)"), IsTrue)
-
-	tk.MustExec("create global binding for select * from t WHERE b=2 AND c=3924541 using select /*+ use_index(@sel_1 test.t idxb) */ * from t WHERE b=2 AND c=3924541")
-	c.Assert(tk.MustUseIndex("SELECT /*+ use_index(@`sel_1` `test`.`t` `idxc`)*/ * FROM `test`.`t` WHERE `b`=2 AND `c`=3924541", "idxb(b)"), IsTrue)
-	c.Assert(tk.MustUseIndex("SELECT /*+ use_index(@`sel_1` `test`.`t` `idxc`)*/ * FROM `t` WHERE `b`=2 AND `c`=3924541", "idxb(b)"), IsTrue)
-
-	// Test for capture baseline
-	s.cleanBindingEnv(tk)
-	stmtsummary.StmtSummaryByDigestMap.Clear()
-	tk.MustExec("set @@tidb_capture_plan_baselines = on")
-	s.domain.BindHandle().CaptureBaselines()
-	c.Assert(tk.Se.Auth(&auth.UserIdentity{Username: "root", Hostname: "%"}, nil, nil), IsTrue)
-	tk.MustExec("select * from t where b=2 and c=213124")
-	tk.MustExec("select * from t where b=2 and c=213124")
-	tk.MustExec("admin capture bindings")
-	rows = tk.MustQuery("show global bindings").Rows()
-	c.Assert(len(rows), Equals, 1)
-	c.Assert(rows[0][0], Equals, "select * from `test` . `t` where `b` = ? and `c` = ?")
-	c.Assert(rows[0][1], Equals, "SELECT /*+ use_index(@`sel_1` `test`.`t` `idxb`)*/ * FROM `test`.`t` WHERE `b` = 2 AND `c` = 213124")
-	tk.MustExec("set @@tidb_capture_plan_baselines = off")
-
-	// Test for evolve baseline
-	s.cleanBindingEnv(tk)
-	tk.MustExec("set @@tidb_evolve_plan_baselines=1")
-	tk.MustExec("create global binding for select * from t WHERE c=3924541 using select /*+ use_index(@sel_1 test.t idxb) */ * from t WHERE c=3924541")
-	rows = tk.MustQuery("show global bindings").Rows()
-	c.Assert(len(rows), Equals, 1)
-	c.Assert(rows[0][0], Equals, "select * from `test` . `t` where `c` = ?")
-	c.Assert(rows[0][1], Equals, "SELECT /*+ use_index(@`sel_1` `test`.`t` `idxb`)*/ * FROM `test`.`t` WHERE `c` = 3924541")
-	tk.MustExec("select /*+ use_index(t idxc)*/ * from t where c=3924541")
-	c.Assert(tk.Se.GetSessionVars().StmtCtx.IndexNames[0], Equals, "t:idxb")
-	tk.MustExec("admin flush bindings")
-	rows = tk.MustQuery("show global bindings").Rows()
-	c.Assert(len(rows), Equals, 2)
-	c.Assert(rows[0][0], Equals, "select * from `test` . `t` where `c` = ?")
-	c.Assert(rows[0][1], Equals, "SELECT /*+ use_index(@`sel_1` `test`.`t` `idxc`)*/ * FROM `test`.`t` WHERE `c` = 3924541")
-	c.Assert(rows[0][3], Equals, "pending verify")
-	tk.MustExec("admin evolve bindings")
-	rows = tk.MustQuery("show global bindings").Rows()
-	c.Assert(len(rows), Equals, 2)
-	c.Assert(rows[0][0], Equals, "select * from `test` . `t` where `c` = ?")
-	c.Assert(rows[0][1], Equals, "SELECT /*+ use_index(@`sel_1` `test`.`t` `idxc`)*/ * FROM `test`.`t` WHERE `c` = 3924541")
-	status := rows[0][3].(string)
-	c.Assert(status == "using" || status == "rejected", IsTrue)
-	tk.MustExec("set @@tidb_evolve_plan_baselines=0")
-}
-
 func (s *testSuite) TestForbidEvolvePlanBaseLinesBeforeGA(c *C) {
 	originalVal := config.CheckTableBeforeDrop
 	config.CheckTableBeforeDrop = false
@@ -1475,24 +1070,6 @@ func (s *testSuite) TestForbidEvolvePlanBaseLinesBeforeGA(c *C) {
 	c.Assert(err, ErrorMatches, "Cannot enable baseline evolution feature, it is not generally available now")
 	err = tk.ExecToErr("admin evolve bindings")
 	c.Assert(err, ErrorMatches, "Cannot enable baseline evolution feature, it is not generally available now")
-}
-
-func (s *testSuite) TestCaptureWithZeroSlowLogThreshold(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	s.cleanBindingEnv(tk)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t(a int)")
-	stmtsummary.StmtSummaryByDigestMap.Clear()
-	c.Assert(tk.Se.Auth(&auth.UserIdentity{Username: "root", Hostname: "%"}, nil, nil), IsTrue)
-	tk.MustExec("set tidb_slow_log_threshold = 0")
-	tk.MustExec("select * from t")
-	tk.MustExec("select * from t")
-	tk.MustExec("set tidb_slow_log_threshold = 300")
-	tk.MustExec("admin capture bindings")
-	rows := tk.MustQuery("show global bindings").Rows()
-	c.Assert(len(rows), Equals, 1)
-	c.Assert(rows[0][0], Equals, "select * from `test` . `t`")
 }
 
 func (s *testSuite) TestExplainTableStmts(c *C) {
@@ -1535,73 +1112,6 @@ func (s *testSuite) TestBindingWithoutCharset(c *C) {
 	c.Assert(len(rows), Equals, 1)
 	c.Assert(rows[0][0], Equals, "select * from `test` . `t` where `a` = ?")
 	c.Assert(rows[0][1], Equals, "SELECT * FROM `test`.`t` WHERE `a` = 'aa'")
-}
-
-func (s *testSuite) TestIssue25505(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	stmtsummary.StmtSummaryByDigestMap.Clear()
-	s.cleanBindingEnv(tk)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	defer func() {
-		tk.MustExec("set tidb_slow_log_threshold = 300")
-	}()
-	tk.MustExec("set tidb_slow_log_threshold = 0")
-	tk.MustExec("create table t (a int(11) default null,b int(11) default null,key b (b),key ba (b))")
-	tk.MustExec("create table t1 (a int(11) default null,b int(11) default null,key idx_ab (a,b),key idx_a (a),key idx_b (b))")
-	tk.MustExec("create table t2 (a int(11) default null,b int(11) default null,key idx_ab (a,b),key idx_a (a),key idx_b (b))")
-	c.Assert(tk.Se.Auth(&auth.UserIdentity{Username: "root", Hostname: "%"}, nil, nil), IsTrue)
-
-	spmMap := map[string]string{}
-	spmMap["with recursive `cte` ( `a` ) as ( select ? union select `a` + ? from `test` . `t1` where `a` < ? ) select * from `cte`"] =
-		"WITH RECURSIVE `cte` (`a`) AS (SELECT 2 UNION SELECT `a` + 1 FROM `test`.`t1` WHERE `a` < 5) SELECT /*+ use_index(@`sel_3` `test`.`t1` `idx_ab`), hash_agg(@`sel_1`)*/ * FROM `cte`"
-	spmMap["with recursive `cte1` ( `a` , `b` ) as ( select * from `test` . `t` where `b` = ? union select `a` + ? , `b` + ? from `cte1` where `a` < ? ) select * from `test` . `t`"] =
-		"WITH RECURSIVE `cte1` (`a`, `b`) AS (SELECT * FROM `test`.`t` WHERE `b` = 1 UNION SELECT `a` + 1,`b` + 1 FROM `cte1` WHERE `a` < 2) SELECT /*+ use_index(@`sel_1` `test`.`t` )*/ * FROM `test`.`t`"
-	spmMap["with `cte1` as ( select * from `test` . `t` ) , `cte2` as ( select ? ) select * from `test` . `t`"] =
-		"WITH `cte1` AS (SELECT * FROM `test`.`t`), `cte2` AS (SELECT 4) SELECT /*+ use_index(@`sel_1` `test`.`t` )*/ * FROM `test`.`t`"
-	spmMap["with `cte` as ( select * from `test` . `t` where `b` = ? ) select * from `test` . `t`"] =
-		"WITH `cte` AS (SELECT * FROM `test`.`t` WHERE `b` = 6) SELECT /*+ use_index(@`sel_1` `test`.`t` )*/ * FROM `test`.`t`"
-	spmMap["with recursive `cte` ( `a` ) as ( select ? union select `a` + ? from `test` . `t1` where `a` > ? ) select * from `cte`"] =
-		"WITH RECURSIVE `cte` (`a`) AS (SELECT 2 UNION SELECT `a` + 1 FROM `test`.`t1` WHERE `a` > 5) SELECT /*+ use_index(@`sel_3` `test`.`t1` `idx_b`), hash_agg(@`sel_1`)*/ * FROM `cte`"
-	spmMap["with `cte` as ( with `cte1` as ( select * from `test` . `t2` where `a` > ? and `b` > ? ) select * from `cte1` ) select * from `cte` join `test` . `t1` on `t1` . `a` = `cte` . `a`"] =
-		"WITH `cte` AS (WITH `cte1` AS (SELECT * FROM `test`.`t2` WHERE `a` > 1 AND `b` > 1) SELECT * FROM `cte1`) SELECT /*+ use_index(@`sel_3` `test`.`t2` `idx_ab`), use_index(@`sel_1` `test`.`t1` `idx_ab`), inl_join(@`sel_1` `test`.`t1`)*/ * FROM `cte` JOIN `test`.`t1` ON `t1`.`a` = `cte`.`a`"
-	spmMap["with `cte` as ( with `cte1` as ( select * from `test` . `t2` where `a` = ? and `b` = ? ) select * from `cte1` ) select * from `cte` join `test` . `t1` on `t1` . `a` = `cte` . `a`"] =
-		"WITH `cte` AS (WITH `cte1` AS (SELECT * FROM `test`.`t2` WHERE `a` = 1 AND `b` = 1) SELECT * FROM `cte1`) SELECT /*+ use_index(@`sel_3` `test`.`t2` `idx_a`), use_index(@`sel_1` `test`.`t1` `idx_a`), inl_join(@`sel_1` `test`.`t1`)*/ * FROM `cte` JOIN `test`.`t1` ON `t1`.`a` = `cte`.`a`"
-
-	tk.MustExec("with cte as (with cte1 as (select /*+use_index(t2 idx_a)*/ * from t2 where a = 1 and b = 1) select * from cte1) select /*+use_index(t1 idx_a)*/ * from cte join t1 on t1.a=cte.a;")
-	tk.MustExec("with cte as (with cte1 as (select /*+use_index(t2 idx_a)*/ * from t2 where a = 1 and b = 1) select * from cte1) select /*+use_index(t1 idx_a)*/ * from cte join t1 on t1.a=cte.a;")
-	tk.MustExec("with cte as (with cte1 as (select /*+use_index(t2 idx_a)*/ * from t2 where a = 1 and b = 1) select * from cte1) select /*+use_index(t1 idx_a)*/ * from cte join t1 on t1.a=cte.a;")
-
-	tk.MustExec("with cte as (with cte1 as (select * from t2 use index(idx_ab) where a > 1 and b > 1) select * from cte1) select /*+use_index(t1 idx_ab)*/ * from cte join t1 on t1.a=cte.a;")
-	tk.MustExec("with cte as (with cte1 as (select * from t2 use index(idx_ab) where a > 1 and b > 1) select * from cte1) select /*+use_index(t1 idx_ab)*/ * from cte join t1 on t1.a=cte.a;")
-	tk.MustExec("with cte as (with cte1 as (select * from t2 use index(idx_ab) where a > 1 and b > 1) select * from cte1) select /*+use_index(t1 idx_ab)*/ * from cte join t1 on t1.a=cte.a;")
-
-	tk.MustExec("WITH RECURSIVE cte(a) AS (SELECT 2 UNION SELECT a+1 FROM t1 use index(idx_ab) WHERE a < 5) SELECT * FROM cte;")
-	tk.MustExec("WITH RECURSIVE cte(a) AS (SELECT 2 UNION SELECT a+1 FROM t1 use index(idx_ab) WHERE a < 5) SELECT * FROM cte;")
-	tk.MustExec("WITH RECURSIVE cte(a) AS (SELECT 2 UNION SELECT a+1 FROM t1 use index(idx_ab) WHERE a < 5) SELECT * FROM cte;")
-
-	tk.MustExec("WITH RECURSIVE cte(a) AS (SELECT 2 UNION SELECT /*+use_index(t1 idx_b)*/  a+1 FROM t1  WHERE a > 5) SELECT * FROM cte;")
-	tk.MustExec("WITH RECURSIVE cte(a) AS (SELECT 2 UNION SELECT /*+use_index(t1 idx_b)*/  a+1 FROM t1  WHERE a > 5) SELECT * FROM cte;")
-	tk.MustExec("WITH RECURSIVE cte(a) AS (SELECT 2 UNION SELECT /*+use_index(t1 idx_b)*/  a+1 FROM t1  WHERE a > 5) SELECT * FROM cte;")
-
-	tk.MustExec("with cte as (select * from t where b=6) select * from t")
-	tk.MustExec("with cte as (select * from t where b=6) select * from t")
-	tk.MustExec("with cte as (select * from t where b=6) select * from t")
-
-	tk.MustExec("with cte1 as (select * from t), cte2 as (select 4) select * from t")
-	tk.MustExec("with cte1 as (select * from t), cte2 as (select 5) select * from t")
-	tk.MustExec("with cte1 as (select * from t), cte2 as (select 6) select * from t")
-
-	tk.MustExec("with recursive cte1(a,b) as (select * from t where b = 1 union select a+1,b+1 from cte1 where a < 2) select * from t")
-	tk.MustExec("with recursive cte1(a,b) as (select * from t where b = 1 union select a+1,b+1 from cte1 where a < 2) select * from t")
-	tk.MustExec("with recursive cte1(a,b) as (select * from t where b = 1 union select a+1,b+1 from cte1 where a < 2) select * from t")
-	tk.MustExec("admin capture bindings")
-	rows := tk.MustQuery("show global bindings").Rows()
-	c.Assert(len(rows), Equals, 7)
-	for _, row := range rows {
-		str := fmt.Sprintf("%s", row[0])
-		c.Assert(row[1], Equals, spmMap[str])
-	}
 }
 
 func (s *testSuite) TestGCBindRecord(c *C) {
@@ -1718,158 +1228,4 @@ func (s *testSerialSuite) TestIssue27422(c *C) {
 	for _, q := range queries {
 		tk.MustGetErrCode(q, errno.ErrOptOnTemporaryTable)
 	}
-}
-
-func (s *testSuite) TestCaptureFilter(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	s.cleanBindingEnv(tk)
-	stmtsummary.StmtSummaryByDigestMap.Clear()
-	tk.MustExec(" set @@tidb_capture_plan_baselines = on")
-	defer func() {
-		tk.MustExec(" set @@tidb_capture_plan_baselines = off")
-	}()
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t(a int)")
-
-	c.Assert(tk.Se.Auth(&auth.UserIdentity{Username: "root", Hostname: "%"}, nil, nil), IsTrue)
-	tk.MustExec("select * from t where a > 10")
-	tk.MustExec("select * from t where a > 10")
-	tk.MustExec("admin capture bindings")
-	rows := tk.MustQuery("show global bindings").Rows()
-	c.Assert(len(rows), Equals, 1)
-	c.Assert(rows[0][0], Equals, "select * from `test` . `t` where `a` > ?")
-
-	// Valid table filter.
-	s.cleanBindingEnv(tk)
-	stmtsummary.StmtSummaryByDigestMap.Clear()
-	tk.MustExec("insert into mysql.capture_plan_baselines_blacklist(filter_type, filter_value) values('table', 'test.t')")
-	tk.MustExec("select * from t where a > 10")
-	tk.MustExec("select * from t where a > 10")
-	tk.MustExec("admin capture bindings")
-	rows = tk.MustQuery("show global bindings").Rows()
-	c.Assert(len(rows), Equals, 0)
-	tk.MustExec("select * from mysql.capture_plan_baselines_blacklist")
-	tk.MustExec("select * from mysql.capture_plan_baselines_blacklist")
-	tk.MustExec("admin capture bindings")
-	rows = tk.MustQuery("show global bindings").Rows()
-	c.Assert(len(rows), Equals, 1)
-	c.Assert(rows[0][0], Equals, "select * from `mysql` . `capture_plan_baselines_blacklist`")
-
-	tk.MustExec("delete from mysql.capture_plan_baselines_blacklist")
-	tk.MustExec("admin capture bindings")
-	rows = tk.MustQuery("show global bindings").Sort().Rows()
-	c.Assert(len(rows), Equals, 2)
-	c.Assert(rows[0][0], Equals, "select * from `mysql` . `capture_plan_baselines_blacklist`")
-	c.Assert(rows[1][0], Equals, "select * from `test` . `t` where `a` > ?")
-
-	// Invalid table filter.
-	s.cleanBindingEnv(tk)
-	stmtsummary.StmtSummaryByDigestMap.Clear()
-	tk.MustExec("insert into mysql.capture_plan_baselines_blacklist(filter_type, filter_value) values('table', 't')")
-	tk.MustExec("select * from t where a > 10")
-	tk.MustExec("select * from t where a > 10")
-	tk.MustExec("admin capture bindings")
-	rows = tk.MustQuery("show global bindings").Rows()
-	c.Assert(len(rows), Equals, 1)
-	c.Assert(rows[0][0], Equals, "select * from `test` . `t` where `a` > ?")
-
-	// Valid database filter.
-	s.cleanBindingEnv(tk)
-	stmtsummary.StmtSummaryByDigestMap.Clear()
-	tk.MustExec("insert into mysql.capture_plan_baselines_blacklist(filter_type, filter_value) values('db', 'mysql')")
-	tk.MustExec("select * from mysql.capture_plan_baselines_blacklist")
-	tk.MustExec("select * from mysql.capture_plan_baselines_blacklist")
-	tk.MustExec("admin capture bindings")
-	rows = tk.MustQuery("show global bindings").Rows()
-	c.Assert(len(rows), Equals, 0)
-	tk.MustExec("select * from t where a > 10")
-	tk.MustExec("select * from t where a > 10")
-	tk.MustExec("admin capture bindings")
-	rows = tk.MustQuery("show global bindings").Rows()
-	c.Assert(len(rows), Equals, 1)
-	c.Assert(rows[0][0], Equals, "select * from `test` . `t` where `a` > ?")
-
-	tk.MustExec("delete from mysql.capture_plan_baselines_blacklist")
-	tk.MustExec("admin capture bindings")
-	rows = tk.MustQuery("show global bindings").Sort().Rows()
-	c.Assert(len(rows), Equals, 2)
-	c.Assert(rows[0][0], Equals, "select * from `mysql` . `capture_plan_baselines_blacklist`")
-	c.Assert(rows[1][0], Equals, "select * from `test` . `t` where `a` > ?")
-
-	// Valid frequency filter.
-	s.cleanBindingEnv(tk)
-	stmtsummary.StmtSummaryByDigestMap.Clear()
-	tk.MustExec("insert into mysql.capture_plan_baselines_blacklist(filter_type, filter_value) values('frequency', '2')")
-	tk.MustExec("select * from t where a > 10")
-	tk.MustExec("select * from t where a > 10")
-	tk.MustExec("admin capture bindings")
-	rows = tk.MustQuery("show global bindings").Rows()
-	c.Assert(len(rows), Equals, 0)
-
-	tk.MustExec("select * from t where a > 10")
-	tk.MustExec("admin capture bindings")
-	rows = tk.MustQuery("show global bindings").Rows()
-	c.Assert(len(rows), Equals, 1)
-	c.Assert(rows[0][0], Equals, "select * from `test` . `t` where `a` > ?")
-	tk.MustExec("delete from mysql.capture_plan_baselines_blacklist")
-
-	// Invalid frequency filter.
-	s.cleanBindingEnv(tk)
-	stmtsummary.StmtSummaryByDigestMap.Clear()
-	tk.MustExec("insert into mysql.capture_plan_baselines_blacklist(filter_type, filter_value) values('frequency', '0')")
-	tk.MustExec("select * from t where a > 10")
-	tk.MustExec("admin capture bindings")
-	rows = tk.MustQuery("show global bindings").Rows()
-	c.Assert(len(rows), Equals, 0)
-
-	tk.MustExec("select * from t where a > 10")
-	tk.MustExec("admin capture bindings")
-	rows = tk.MustQuery("show global bindings").Rows()
-	c.Assert(len(rows), Equals, 1)
-	c.Assert(rows[0][0], Equals, "select * from `test` . `t` where `a` > ?")
-	tk.MustExec("delete from mysql.capture_plan_baselines_blacklist")
-
-	// Invalid filter type.
-	s.cleanBindingEnv(tk)
-	stmtsummary.StmtSummaryByDigestMap.Clear()
-	tk.MustExec("insert into mysql.capture_plan_baselines_blacklist(filter_type, filter_value) values('unknown', 'xx')")
-	tk.MustExec("select * from t where a > 10")
-	tk.MustExec("select * from t where a > 10")
-	tk.MustExec("admin capture bindings")
-	rows = tk.MustQuery("show global bindings").Rows()
-	c.Assert(len(rows), Equals, 1)
-	c.Assert(rows[0][0], Equals, "select * from `test` . `t` where `a` > ?")
-	tk.MustExec("delete from mysql.capture_plan_baselines_blacklist")
-
-	// Case sensitivity.
-	s.cleanBindingEnv(tk)
-	stmtsummary.StmtSummaryByDigestMap.Clear()
-	tk.MustExec("insert into mysql.capture_plan_baselines_blacklist(filter_type, filter_value) values('tABle', 'tESt.T')")
-	tk.MustExec("select * from t where a > 10")
-	tk.MustExec("select * from t where a > 10")
-	tk.MustExec("admin capture bindings")
-	rows = tk.MustQuery("show global bindings").Rows()
-	c.Assert(len(rows), Equals, 0)
-
-	tk.MustExec("delete from mysql.capture_plan_baselines_blacklist")
-	tk.MustExec("admin capture bindings")
-	rows = tk.MustQuery("show global bindings").Sort().Rows()
-	c.Assert(len(rows), Equals, 1)
-	c.Assert(rows[0][0], Equals, "select * from `test` . `t` where `a` > ?")
-
-	s.cleanBindingEnv(tk)
-	stmtsummary.StmtSummaryByDigestMap.Clear()
-	tk.MustExec("insert into mysql.capture_plan_baselines_blacklist(filter_type, filter_value) values('Db', 'mySQl')")
-	tk.MustExec("select * from mysql.capture_plan_baselines_blacklist")
-	tk.MustExec("select * from mysql.capture_plan_baselines_blacklist")
-	tk.MustExec("admin capture bindings")
-	rows = tk.MustQuery("show global bindings").Rows()
-	c.Assert(len(rows), Equals, 0)
-
-	tk.MustExec("delete from mysql.capture_plan_baselines_blacklist")
-	tk.MustExec("admin capture bindings")
-	rows = tk.MustQuery("show global bindings").Sort().Rows()
-	c.Assert(len(rows), Equals, 1)
-	c.Assert(rows[0][0], Equals, "select * from `mysql` . `capture_plan_baselines_blacklist`")
 }
